netlink: do not alter set element width
Consider the following rulset

table bridge t {
        set nodhcpvlan {
                typeof vlan id
                elements = { 1 }
        }

        chain c1 {
                vlan id != @nodhcpvlan vlan type arp counter packets 0 bytes 0 jump c2
                vlan id != @nodhcpvlan vlan type ip counter packets 0 bytes 0 jump c2
        }

        chain c2 {
        }
}

This results for nft list ruleset in
  nft: netlink_delinearize.c:1945: binop_adjust_one: Assertion `value->len >= binop->right->len' failed.

This is due to binop_adjust_one setting value->len to left->len, which
is shorther than right->len.

Additionally, it does not seem correct to alter set elements from parsing a rule,
so remove that part all together.

Signed-off-by: Michael Braun <michael-dev@fami-braun.de>